+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Added `sx-extract` archive extraction/dumping tool ([#114])
+
 ### Removed
 - Drop support for Python 2.7 ([#115])
 
@@ -161,6 +164,7 @@ Initial release
 [#102]: https://github.com/JonathonReinhart/staticx/pull/102
 [#111]: https://github.com/JonathonReinhart/staticx/pull/111
 [#112]: https://github.com/JonathonReinhart/staticx/pull/112
+[#114]: https://github.com/JonathonReinhart/staticx/pull/114
 [#115]: https://github.com/JonathonReinhart/staticx/pull/115
 [#118]: https://github.com/JonathonReinhart/staticx/pull/118
 [#120]: https://github.com/JonathonReinhart/staticx/pull/120